将本地缓存文件的功能拆分为函数
